backend/requirements.txt
fastapi
uvicorn
sqlalchemy
streamlit
requests
numpy
pandas
plotly
python-dotenv
psycopg2-binary